Bug fixes. Preparing to see if there is another solution to the circular references in the proxyowner/proxy pattern.

+++ src/plonk/core/plonk_SmartPointer.h
@@ -59,7 +59,7 @@ public:
 	/// @{
 	
 	SmartPointer (const bool allocateWeakPointer = true) throw();
-	virtual ~SmartPointer();
+    virtual ~SmartPointer();
     
     static void* operator new (size_t size); 
     static void operator delete (void* ptr);
@@ -79,24 +79,69 @@ public:
 	/// @name Miscellaneous
 	/// @{
 	
-	Int getRefCount() const throw()	{ return refCount.getValueUnchecked(); }
-//	bool isActive() const throw()       { return active != 0; }
-    void update() throw() { }
+	Long getRefCount() const throw()	{ return refCount.getValueUnchecked(); }
+    void update() throw()               { }
     void* getWeak() const throw()       { return weakPointer; }
-        
+    
 	/// @} <!-- end Miscellaneous -->
-		
+    
 protected:
     virtual void destroy();
     
-    //Int refCount;
-    AtomicValue<Int> refCount;
+    AtomicValue<Long> refCount;
     void *weakPointer;
-//	unsigned int active;
 	
 private:
 	SmartPointer (const SmartPointer&);
     const SmartPointer& operator= (const SmartPointer&);
 };
+
+//class SmartPointer
+//{
+//public:
+//	
+//	/// @name Construction and destruction
+//	/// @{
+//	
+//	SmartPointer (const bool allocateWeakPointer = true) throw();
+//    virtual ~SmartPointer();
+//
+//    static void* operator new (size_t size); 
+//    static void operator delete (void* ptr);
+//    
+//	void incrementRefCount() throw();
+//    bool decrementRefCount() throw(); 
+//    
+//    /** To check for mutual-only references remaining.
+//     Most SmartPointer subclass don't need this but some which naturally form
+//     circular references may need to detect this state and release themselves.
+//     @see ProxyOwnerChannelInternal, ProxyChannelInternal
+//     */
+//    virtual bool deleteIfOnlyMutualReferencesRemain() throw() { return false; } 
+//	
+//	/// @} <!-- end Construction and destruction -->
+//	
+//	/// @name Miscellaneous
+//	/// @{
+//	
+//	Long getRefCount() const throw()	{ return refCount.getValueUnchecked(); }
+////	bool isActive() const throw()       { return active != 0; }
+//    void update() throw() { }
+//    void* getWeak() const throw()       { return weakPointer; }
+//        
+//	/// @} <!-- end Miscellaneous -->
+//		
+//protected:
+//    virtual void destroy();
+//    
+//    //Int refCount;
+//    AtomicValue<Long> refCount;
+//    void *weakPointer;
+////	unsigned int active;
+//	
+//private:
+//	SmartPointer (const SmartPointer&);
+//    const SmartPointer& operator= (const SmartPointer&);
+//};
 
 #endif // PLONK_SMARTPOINTER_H
